# support parsing headers with fast-mail-parser