@bruth 2023-05-19
@RouquinBlanc Thanks for the analysis and contribution!

It looks like the CI failed due to [linting](https://app.travis-ci.com/github/nats-io/nats.py/jobs/597162669#L313). Do you mind reworking that list comprehension to reduce the line length?