@wallyqs 2023-06-01
Thanks for the contribution! Merged with some fixes at: https://github.com/nats-io/nats.py/commit/ad5f2bf9515216951448d1a85d0a371a57a83f3e